WebCore/Sources.txt: Remove #if ENABLE_WIRELESS_PLAYBACK_TARGET https://bugs.webkit.org/show_bug.cgi?id=213945
Reviewed by Sam Weinig.

* CMakeLists.txt: Moved items from the pictureinpicture and remoteplayback module here
from PlatformMac.cmake. Even if these are currently only implemented on Mac, they are
not intrinsically Mac-specific, so they belong here.
* PlatformMac.cmake: Ditto.

* Sources.txt: Remove #if ENABLE_WIRELESS_PLAYBACK_TARGET and sort files in with the others.
The source files themselves have conditionals, we don't need conditionals in Sources.txt,
and we are moving towards removing the capability for #if ENABLE in the Sources.txt file.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263918 268f45cc-cd09-0410-ab3c-d52691b4dbfc