[JSC] Promise should check whether a user-provided function is set by using `@isUndefinedOrNull` https://bugs.webkit.org/show_bug.cgi?id=213951
Reviewed by Ross Kirsling.

JSTests:

* stress/masquerader-as-function-in-promise.js: Added.

Source/JavaScriptCore:

If a user-provided function is masquerade-as-undefined value, `if (!xxx.@onRejected)` returns wrong
value since this function object is considered as undefined in this context. We should use `@isUndefinedOrNull`
here instead since this if-branch is checking whether this property is null/undefined actually.
And `if (@isUndefinedOrNull(...))` is efficient since we have `jundefined_or_null` / `jnundefined_or_null` bytecodes.

* builtins/PromiseOperations.js:
(globalPrivate.promiseReactionJob):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263930 268f45cc-cd09-0410-ab3c-d52691b4dbfc