[JSC] Fix btjs by recovering CallFrame::describeFrame https://bugs.webkit.org/show_bug.cgi?id=214055
Reviewed by Mark Lam.

While CallFrame::describeFrame is not used in WebKit tree, it is used in LLDB btjs which is invoked from python.
So we need to keep it. We also use std::call_once based buffer allocation instead of `static char buffer[...]`
to avoid spreading debug-use-only buffer in BSS segment.

* interpreter/CallFrame.cpp:
(JSC::CallFrame::describeFrame):
* interpreter/CallFrame.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264051 268f45cc-cd09-0410-ab3c-d52691b4dbfc