[JHBuild] Add support for using a minimal moduleset https://bugs.webkit.org/show_bug.cgi?id=213614
Reviewed by Carlos Garcia Campos.

This patch introduces a way of specifying a moduleset other than the default
via the environment variable WEBKIT_JHBUILD_MODULESET and adds a minimal moduleset
that allows building WebKit with it. This minimal moduleset includes the libraries
needed for building WebKit on Ubuntu-18.04, and those libraries are only build if
needed since they include the <pkg-config> entries that tells JHBuild to not build
them if those are already installed system wide. So this minimal moduleset should
work also on newer versions of Ubuntu or even other distributions.

Currently a recipe for newer libsoup than the one shipped by Ubuntu-18.04 is included
since we need this to have support for SameSite cookie support.
The minimal moduleset is shared between WPE and GTK to reduce code duplication.

To use this, you have to export on the build environment:
WEBKIT_JHBUILD=1
WEBKIT_JHBUILD_MODULESET=minimal

* Scripts/update-webkit-libs-jhbuild:
(getJhbuildIncludedFilePaths):
(jhbuildModulesetNameChanged):
(jhbuildConfigurationChanged):
(saveJhbuildMd5):
(saveJhbuildModulesetName):
(saveJhbuildConfig):
(deleteJhbuildMd5):
(deleteJhbuildModulesetName):
(deleteJhbuildConfig):
* Scripts/webkitdirs.pm:
(getJhbuildModulesetName):
* gtk/jhbuild-minimal.modules: Added.
* jhbuild/jhbuild-minimal.modules: Added.
* jhbuild/jhbuildrc_common.py:
(init):
* jhbuild/patches/libsoup-lower-glib-dependency-to-2.38.patch: Added.
* wpe/jhbuild-minimal.modules: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264092 268f45cc-cd09-0410-ab3c-d52691b4dbfc