[WPE] Stick to default Glib source priorities in HeadlessViewBackend, WindowViewBackend https://bugs.webkit.org/show_bug.cgi?id=214249
Reviewed by Carlos Garcia Campos.

The -70 GLib source priority was a magic baseline used to align other
GLib source priorities throughout the WPE stack. We moved to the
G_PRIORITY_DEFAULT value since then, so that's the value that should
be used in the two in-tree view backend implementations as well.

* wpe/backends/HeadlessViewBackend.cpp:
(WPEToolingBackends::HeadlessViewBackend::HeadlessViewBackend):
* wpe/backends/WindowViewBackend.cpp:
(WPEToolingBackends::WindowViewBackend::WindowViewBackend):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264295 268f45cc-cd09-0410-ab3c-d52691b4dbfc