REGRESSION(r262341) URL::createCFURL should produce a CFURL that uses UTF-8 to decode its percent-encoded sequences https://bugs.webkit.org/show_bug.cgi?id=214314 <rdar://problem/65079249>
Patch by Alex Christensen <achristensen@webkit.org> on 2020-07-14
Reviewed by Darin Adler.

Source/WTF:

r262341 made it so we usually pass kCFStringEncodingISOLatin1 into CFURLCreateAbsoluteURLWithBytes when creating a CFURLRef from a WTF::URL.
This is correct for the interpretation of the bytes to create a CFStringRef, and that is a change we want to make.
The encoding, however, is also stored with the CFURL and used later when interpreting percent-encoded sequences.
We want to use kCFStringEncodingUTF8 to make the interpretation of percent-encoded sequences the same as it used to be.
To avoid making a separate CString most of the time, use characters8() only for ASCII-only URLs, where UTF-8 and Latin1 are the same.
For all other URLs, we have to make a CString containing the UTF-8 representation of the string to get the percent-encoded sequences interpreted the same.

* wtf/cf/URLCF.cpp:
(WTF::URL::createCFURL const):
* wtf/cocoa/URLCocoa.mm:
(WTF::URL::createCFURL const):

Tools:

* TestWebKitAPI/Tests/WTF/cocoa/URLExtras.mm:
Add a direct test and update the Mojave test expectations for the test introduced in r262341 with the change that introduced this regression.
* TestWebKitAPI/Tests/WebKitCocoa/LoadInvalidURLRequest.mm:
Update a test from r263475 which fixed another regression from r262341.
The test verified that nothing crashed or timed out, and that is still the case even with this intentional change in behavior.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264382 268f45cc-cd09-0410-ab3c-d52691b4dbfc