[WK2] Make establishing a connection between the WebProcess and the Network process more robust https://bugs.webkit.org/show_bug.cgi?id=214307 <rdar://problem/64592340>
Reviewed by Alex Christensen.

Make establishing a connection between the WebProcess and the Network process more robust.
When the network process is crashy, it is not uncommon for the WebProcess to fail to establish
a connection to the network process. Previously, we would try twice and then give up by calling
CRASH(). I have tweaked the logic so that we now try 10 times and wait 100 ms between each
attempt to make a best effort at establishing the connection. I also updated the logic to exit
cleanly instead of crashing if the IPC connection to the UIProcess becomes invalid (i.e. the
UIProcess has severed the connection because this WebProcess is no longer needed).

* WebProcess/WebProcess.cpp:
(WebKit::getNetworkProcessConnection):
(WebKit::WebProcess::ensureNetworkProcessConnection):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264408 268f45cc-cd09-0410-ab3c-d52691b4dbfc