REGRESSION(r205826): narrowNoBreakSpace (U+202F) has zero width, regardless of font https://bugs.webkit.org/show_bug.cgi?id=214626 <rdar://problem/65193790>
Reviewed by Timothy Hatcher.

Source/WebCore:

There's no reason to overwrite narrowNoBreakSpace with zeroWidthSpace.

This is becoming much more prominent now because a lot of the French unit formatting data in CLDR - and hence in e.g. NSMeasurementFormatter - uses 202F between numeric value and unit, e.g. "37 km"

It works in native content on macOS, and in all other browsers on macOS, but not in WebKit.

Test: fast/text/narrowNoBreakSpace.html

* platform/graphics/Font.cpp:
(WebCore::overrideControlCharacters):

LayoutTests:

* fast/text/narrowNoBreakSpace-expected-mismatch.html: Added.
* fast/text/narrowNoBreakSpace.html: Added.
* fast/text/narrow-non-breaking-space-expected-mismatch.html: Renamed from fast/text/narrow-non-breaking-space-expected.html.
* fast/text/narrow-non-breaking-space.html: Made the test more robust.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264714 268f45cc-cd09-0410-ab3c-d52691b4dbfc