[webkitcorepy] Make scripts called with sudo use a userspace autoinstall https://bugs.webkit.org/show_bug.cgi?id=214981 <rdar://problem/66342996>
Reviewed by Dewei Zhu.

Sudo can bring havoc to the autoinstaller, because the directories it creates will belong
to the root user. We should ensure that the directories created by a sudo script call
are owned by the user that called sudo, when possible.

* Scripts/libraries/webkitcorepy/webkitcorepy/__init__.py:
* Scripts/libraries/webkitcorepy/webkitcorepy/autoinstall.py:
(Package.install): Ensure that installed packages are not owned by root, if possible.
(AutoInstall.userspace_should_own): Ensure that a directory or file is not owned by root, if possible.
(AutoInstall.set_directory): Ensure that the directories containing auto-installed packages are not
owned by root, if possible.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@265123 268f45cc-cd09-0410-ab3c-d52691b4dbfc