[Mac] AudioSessionRoutingArbitrator causes a launch time regression checking for CoreAudio muted state https://bugs.webkit.org/show_bug.cgi?id=214993
Reviewed by Eric Carlson.

The initial query of isMuted() is only there to tell whether, when we get a notification that the mute state
changed, whether our internal state is dirty and we need to fire a notification. Instead, replace the bool member
with an Optional<bool>, so we know we need to fire a changed notification whenever the first mute state change
comes in.

* platform/audio/mac/AudioSessionMac.mm:
(WebCore::AudioSession::AudioSession):
(WebCore::AudioSession::handleMutedStateChange):
(WebCore::AudioSessionPrivate::AudioSessionPrivate): Deleted.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@265173 268f45cc-cd09-0410-ab3c-d52691b4dbfc