Add the ability of comparing the accessibility tree with isolated tree mode on and off. https://bugs.webkit.org/show_bug.cgi?id=215000
Reviewed by Chris Fleizach.

Source/WebCore:

Test: accessibility/mac/isolated-tree-mode-on-off.html

Added [WebAccessibilityObjectWrapper isIsolatedObject] for testing
purposes, so that layout tests can determine whether an object is
isolated or not. It should not be used by real clients since there
should be no difference in behavior of a WebAccessibilityObjectWrapper
from a client point of view, regardless the underlying object being
isolated or not.

* accessibility/mac/WebAccessibilityObjectWrapperBase.mm:
(-[WebAccessibilityObjectWrapperBase isIsolatedObject]):

Tools:

* WebKitTestRunner/InjectedBundle/AccessibilityController.cpp:
(WTR::AccessibilityController::setIsolatedTreeMode):
(WTR::AccessibilityController::setAccessibilityIsolatedTreeMode): Renamed setIsolatedTreeMode.
* WebKitTestRunner/InjectedBundle/AccessibilityController.h:
* WebKitTestRunner/InjectedBundle/AccessibilityUIElement.h:
* WebKitTestRunner/InjectedBundle/Bindings/AccessibilityController.idl:
Exposes setIsolatedTreeMode to JavaScript code so that it can be turned
on and off in a layout tests.
* WebKitTestRunner/InjectedBundle/Bindings/AccessibilityUIElement.idl:
Exposes isIsolatedObject to JavaScript code so that layout tests can
check whether the underlying accessibility object is an AXIsolatedObject
or an AXObject.
* WebKitTestRunner/InjectedBundle/InjectedBundle.cpp:
(WTR::InjectedBundle::beginTesting): Rename.
* WebKitTestRunner/InjectedBundle/mac/AccessibilityUIElementMac.mm:
(WTR::AccessibilityUIElement::isIsolatedObject const):

LayoutTests:

* accessibility/mac/isolated-tree-mode-on-off-expected.txt: Added.
* accessibility/mac/isolated-tree-mode-on-off.html: Added.
* resources/accessibility-helper.js:
(dumpAccessibilityTree): In addition to dumping the accessibility tree
into an element with id="tree", it returns a string representation of
the accessibility tree.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@265248 268f45cc-cd09-0410-ab3c-d52691b4dbfc